requirements.txt
beautifulsoup4==4.12.2
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
Pillow==9.5.0
platformdirs==4.1.0
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
xml2epub==2.6.3
